Intersection Now Working
Need to work on lighting now

+++ util/Basic/src/Light.cpp
@@ -25,16 +25,16 @@ void PointLight::shade(Ray& _ray){
 		float k = _ray.m_pIntersectionProperties->m_Normal.dot(-light_direction) * 2;
 		Vector4f reflect_direction = light_direction - k * _ray.m_pIntersectionProperties->m_Normal;
         reflect_direction.Normalize();
-		double diffuse = std::max(0.0,(double)_ray.m_pIntersectionProperties->m_Normal.dot(light_direction));
-		double specular = std::pow(std::max(0.0,(double)ray_direction.dot(-reflect_direction)),
+		float diffuse = std::max(0.0,(double)_ray.m_pIntersectionProperties->m_Normal.dot(light_direction));
+		float specular = std::pow(std::max(0.0,(double)ray_direction.dot(-reflect_direction)),
 				_ray.m_pIntersectionProperties->m_Material->m_fSpecularWeight);
-		double red = this->m_LightingAttribute.m_AmbientColor[0]*_ray.m_pIntersectionProperties->m_Material->m_AmbientColor[0]+
+		float red = this->m_LightingAttribute.m_AmbientColor[0]*_ray.m_pIntersectionProperties->m_Material->m_AmbientColor[0]+
 				this->m_LightingAttribute.m_DefuseColor[0]*_ray.m_pIntersectionProperties->m_Material->m_DefuseColor[0]*diffuse+
 				this->m_LightingAttribute.m_SpecularColor[0]*_ray.m_pIntersectionProperties->m_Material->m_SpecularColor[0]*specular;
-		double green = this->m_LightingAttribute.m_AmbientColor[1]*_ray.m_pIntersectionProperties->m_Material->m_AmbientColor[1]+
+		float green = this->m_LightingAttribute.m_AmbientColor[1]*_ray.m_pIntersectionProperties->m_Material->m_AmbientColor[1]+
 						this->m_LightingAttribute.m_DefuseColor[1]*_ray.m_pIntersectionProperties->m_Material->m_DefuseColor[1]*diffuse+
 						this->m_LightingAttribute.m_SpecularColor[1]*_ray.m_pIntersectionProperties->m_Material->m_SpecularColor[1]*specular;
-		double blue = this->m_LightingAttribute.m_AmbientColor[2]*_ray.m_pIntersectionProperties->m_Material->m_AmbientColor[2]+
+		float blue = this->m_LightingAttribute.m_AmbientColor[2]*_ray.m_pIntersectionProperties->m_Material->m_AmbientColor[2]+
 						this->m_LightingAttribute.m_DefuseColor[2]*_ray.m_pIntersectionProperties->m_Material->m_DefuseColor[2]*diffuse+
 						this->m_LightingAttribute.m_SpecularColor[2]*_ray.m_pIntersectionProperties->m_Material->m_SpecularColor[2]*specular;
 		_ray.m_color[0] *= red;
